ios, ci: enforce JS↔Swift parser parity with a shared-fixture swiftc harness
Extract PollTextParser out of AppDelegate.swift into its own pure-Foundation
PollTextParser.swift (4-entry pbxproj patch) so scripts/ios/test-parser.sh can
swiftc-compile it standalone with scripts/ios/parser-harness/main.swift, which
reads the SAME tests/fixtures/poll-parse-cases.json the JS test asserts and
checks PollTextParser.decide case-by-case. ios-build.yml runs it on the Mac
runner before archiving, so a drifting Swift port now fails the iOS build.
Closes docs/poll-textbox-followups.md TODO 1(a) (supersedes the XCTest-target
idea — no test target / simulator needed).

Also mark the Latest-Tier Rollout steps 5b + 6 and the notification-tick
deploy caveat as verified-done in CLAUDE.md (confirmed 2026-06-11 on both
droplets: release-event deploy in the prod webhook log, empty
/etc/droplet-label, per-minute tick crontab + INTERNAL_TICK_SECRET present).
